fix [integration/vllm]: fix kv cache allocation bug; move autopatch inside vllm-specific subdir
